Make sectionID a long since the range is larger than an integer

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/J2735VehicleID.java
@@ -1,12 +1,14 @@
 package us.dot.its.jpo.ode.plugin.j2735;
 
+import java.math.BigInteger;
+
 import us.dot.its.jpo.ode.plugin.asn1.Asn1Object;
 
 public class J2735VehicleID extends Asn1Object {
 
     private static final long serialVersionUID = 1L;
     private String entityID;
-    private Integer stationID;
+    private Long stationID;
 
     public String getEntityID() {
         return entityID;
@@ -16,11 +18,11 @@ public class J2735VehicleID extends Asn1Object {
         this.entityID = entityID;
     }
 
-    public Integer getStationID() {
+    public Long getStationID() {
         return stationID;
     }
 
-    public void setStationID(Integer stationID) {
+    public void setStationID(Long stationID) {
         this.stationID = stationID;
     }
 }
