+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/SignalRequesterInfoBuilder.java
@@ -28,7 +28,7 @@ public class SignalRequesterInfoBuilder {
 			JsonNode stationID = id.get("stationID");
 			if(stationID != null)
 			{
-				objVehicleId.setStationID(stationID.asInt());
+				objVehicleId.setStationID(stationID.asLong());
 			}
 
 			signalRequesterInfo.setId(objVehicleId);
